Chapter 3-Programming Project [P-3.57].py
The provided code performs the following tasks:

It imports necessary libraries:

time: For measuring the execution time of sorting.
numpy (as np): For generating random data and numerical operations.
matplotlib.pylab (as plt): For creating a plot to visualize the results.
It defines a function do_sorted(n):

This function takes an integer n as input.
It generates an array A of n random integers between 0 and 1 million (bignum).
It sorts the array A using the sorted function (which sorts it in ascending order).
It sets up variables to collect data for plotting:

plot_n: An empty list to store the input sizes (n values).
plot_walltime: An empty list to store the execution times for sorting.
It enters a loop that iterates over a range of input sizes (j values):

n is set to the integer value of j.
It records the current time (start) before calling the do_sorted(n) function.
It records the time again after the sorting is completed (end).
It calculates and prints the elapsed time for sorting an array of size n.
It appends the input size n and the corresponding wall time to the plot_n and plot_walltime lists.
After the loop finishes, it uses matplotlib to create a plot:

The x-axis represents the input size (n), ranging from 1000 to 100,000.
The y-axis represents the wall time (execution time) taken to sort the array of that size.
It plots the data points (plot_n vs. plot_walltime) as markers ('o') connected by lines ('-').
Finally, it displays the plot using plt.show().